Fleshed out the d1_libclient_java CNode to support getFormat():
Modified getFormat() to accept the format id as a parameter, and to deserialize the input stream received from the CN service.  
Also added InsufficientResources exception handling to getFormat() and listFormats().

In D1Node, added the deserializeObjectFormat() method which returns the ObjectFormat instance for the given identifier.  This may be useful to MNode as well.