Modify ProvResourceMapBuilder.insertRelationship() to take a list of RDFNode objects as the objects for the triple statements.  This will allow us to make statements using literal types, and not just Resources.
Update DataPackage to store triples in the tripleMap using RDFNode objects as well to conform to the change above. Overload insertRelationship() to take a single URI object to accommodate languages that don't support generics typing (like Matlab).  Change the overloaded insertRelationship() method that took a String object to now be an Object literal.  This is unimplemented yet - work in progress.
refs #7154