This is a significant change in DataPackage to redesign how we store the triple statements.  After having a design discussion with Ben regarding the limitations of using a Map<Identifier, Map<Identifier, List<Identifier>>> structure, change the internal storage to be the ResourceMap itself.  This removes the need of having 1) a metadataMap, and 2) a tripleMap.  The tripleMap itself was the most problematic, in that it was impossible to ensure insertion order when creating a resource map from the LinkedHashMap because of repeating Predicate keys.  Both maps are replaced by the resourceMap, which is basically just a list of RDF statements anyway.
In support of this, change the constructor to create a base resource map from ResourceMapFactory, and modify insertRelationship(Identifier, List<Identifier>) to build the ResourceMap instance up rather than the metadataMap.  It now aggregates resources into the map, and describes them with other calls to insertRelationship().  Add a private aggregate() method to help with this.  change getMap() to getResourceMap(), and change setMetadataMap() to setResourceMap().  Keep getMetadataMap() since it's helpful in serializing/deserializing using ResourceMapFactory.  Change getDocumentedBy() to search the resource map rather than the old metadata map whaen returning a metadata identifier.

lastly, finish out the version of insertRelationship() that allows for statements with literals as the object.

refs #7154